[breaking]: improving data model: recordings is referenced, (not embedded anymore)
