+++ app/server/models/vocabLog.js
@@ -1,25 +1,16 @@
 const mongoose = require('mongoose');
 
-const recordingSchema = new mongoose.Schema(
-  {
-    binaryAudioData: {
-      type: Buffer, // Store recording data as a Buffer
-      required: true,
-      subtype: 0x02, // Set the subtype to "audio"
-    },
-    intelligibilityScore: Boolean,
-    // TODO: download mongoose-float npm package to use Float type
-    is_recognized: Number,
-  },
-  { timestamps: true }
-);
-
 const vocabLogSchema = new mongoose.Schema(
   {
     spokenWords: [
       {
         word: String,
-        recordings: [recordingSchema],
+        recordings: [
+          {
+            type: mongoose.Schema.Types.ObjectId,
+            ref: 'Recording',
+          },
+        ],
         peerComparisonScore: {
           type: Number,
           enum: [1, 2, 3, 4, 5],
@@ -47,6 +38,5 @@ vocabLogSchema.set('toJSON', {
 });
 
 const VocabLog = mongoose.model('VocabLog', vocabLogSchema);
-const Recording = mongoose.model('Recording', recordingSchema);
 
-module.exports = { VocabLog, Recording };
+module.exports = VocabLog;
